Add Sentry and fix settings
