--- {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/manage.py
@@ -1,12 +1,0 @@
-import os
-import sys
-
-
-if __name__ == '__main__':
-    os.environ.setdefault(
-        'TORNADO_SETTINGS_MODULE', 'settings.development'
-    )
-
-    from contrib.execute_from_command_line import execute_from_command_line
-
-    execute_from_command_line(sys.argv)
