+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/__init__.py
@@ -1,7 +1,6 @@
 import os
 import importlib 
 
-print(os.getenv('TORNADO_SETTINGS_MOULE'))
 
 # import all settings
 settings = importlib.import_module(os.getenv('TORNADO_SETTINGS_MODULE'))
