+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/base.py
@@ -1,18 +1,12 @@
 import os
 
-from dotenv import load_dotenv
-
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 PROJECT_DIR = os.path.dirname(BASE_DIR)
 
-# load environment config
-dotenv_path = os.path.join(BASE_DIR, '.env')
-load_dotenv(dotenv_path)
-
 SECRET_KEY = os.getenv('SECRET_KEY')
 
-DEBUG = os.getenv('DEBUG', True)
+DEBUG = bool(os.getenv('DEBUG', True))
 
 GZIP = True
 
@@ -20,7 +14,7 @@ SQL_ECHO = True
 
 DATABASES = {
     'default': {
-        'ENGINE': 'mssql+pymssql',
+        'ENGINE': 'sqlite',
         'HOST': os.getenv('DEFAULT_HOST'),
         'NAME': os.getenv('DEFAULT_NAME'),
         'USER': os.getenv('DEFAULT_USER'),
@@ -33,3 +27,5 @@ TEMPLATE_ROOT = os.path.join(os.path.dirname(BASE_DIR), 'templates')
 
 STATIC_URL = '/static/'
 STATIC_ROOT = os.path.join(os.path.dirname(BASE_DIR), 'static')
+
+SENTRY_DSN = os.getenv('SENTRY_DSN')
