+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/settings/development.py
@@ -1,9 +1,1 @@
 from .base import *
-
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'sqlite',
-        'NAME': ''
-    }
-}
